Allow thread to sleep for 1 sec before exporting analytics
This is a workaround. The trigger code underneath takes almost 2ms to 1s to run the logic before it saves the row on to the DB. Some rows might not get exported if the export begins before the completion. Until we find a way to dynamically know the time when a row actually get saved on to the DB, we can go ahead with this approach;