build(Dhis2Verifier): ensure the version is repo-wide
This is a decision we may have to revisit in the future. Right now, we
cannot release independent versions of tooling within the repository.
This is because the Maven packages depends on versioned POMs, and we
currently are not publishing these POMs anywhere. So there's no way for
us to independently pin versions on anything since we would not be able
to find said versions anywhere. Difficulty with this is, releasing a
breaking change in one tool means a breaking change in every tool; which
technically isn't true, and isn't a good experience for the community.